pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wordpress-mcp-server"
version = "1.0.0"
description = "A Model Context Protocol (MCP) server for WordPress blog management with Claude AI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "WordPress MCP Contributors", email = "contact@example.com"}
]
maintainers = [
    {name = "WordPress MCP Contributors", email = "contact@example.com"}
]
keywords = ["mcp", "wordpress", "claude", "ai", "blog", "automation", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: Content Management System",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: HTML",
]
dependencies = [
    "aiohttp>=3.9.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "build>=0.10.0",
    "twine>=4.0.0"
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "aioresponses>=0.7.0"
]

[project.urls]
Homepage = "https://github.com/yourusername/wordpress-mcp-server"
Documentation = "https://github.com/yourusername/wordpress-mcp-server#readme"
Repository = "https://github.com/yourusername/wordpress-mcp-server.git"
"Bug Tracker" = "https://github.com/yourusername/wordpress-mcp-server/issues"
Changelog = "https://github.com/yourusername/wordpress-mcp-server/blob/main/CHANGELOG.md"

[project.scripts]
wordpress-mcp-server = "wordpress_mcp_server.cli:main"
wp-mcp = "wordpress_mcp_server.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wordpress_mcp_server = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]